Rename project to react-native-image-utils
Namespace under com.appfolio.react.imageutils for Android and use the AE prefix
for iOS (RCT has special meaning).

Used react-native-create-library as the starting template and copied over code
and configuration as appropriate.
