+++ .gitignore
@@ -1,6 +1,14 @@
+
 # OSX
 #
 .DS_Store
+
+# node.js
+#
+node_modules/
+npm-debug.log
+yarn-error.log
+
 
 # Xcode
 #
@@ -22,6 +30,7 @@ DerivedData
 *.xcuserstate
 project.xcworkspace
 
+
 # Android/IntelliJ
 #
 build/
@@ -30,28 +39,7 @@ build/
 local.properties
 *.iml
 
-# node.js
-#
-node_modules/
-npm-debug.log
-
 # BUCK
 buck-out/
 \.buckd/
-android/app/libs
 *.keystore
-
-# fastlane
-#
-# It is recommended to not store the screenshots in the git repo. Instead, use fastlane to re-generate the
-# screenshots whenever they are needed.
-# For more information about the recommended setup visit:
-# https://github.com/fastlane/fastlane/blob/master/fastlane/docs/Gitignore.md
-
-fastlane/report.xml
-fastlane/Preview.html
-fastlane/screenshots
-
-
-# APK
-android/app/*.apk
