+++ LICENSE
@@ -1,5 +1,6 @@
 The MIT License (MIT)
 
+Copyright (c) 2018 AppFolio, Inc.
 Copyright (c) 2015 Florian Rival and Alexandre Moureaux
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
