+++ README.md
@@ -1,46 +1,46 @@
-# React Native Image Resizer
+# React Native Image Utils
 
 A React Native module that can create scaled versions of local images (also supports the assets library on iOS).
 
-## Setup
+## Getting started
 
-Install the package:
+`$ npm install react-native-image-utils --save`
 
-* 😻 React Native >= 0.40
-```
-npm install --save react-native-image-resizer
-react-native link react-native-image-resizer
-```
+### Mostly automatic installation
 
-> ⚠️ **Heads up, breaking change!** If you're upgrading *react-native-image-resizer* to version 1.0.0, please note that the response of `createResizedImage` changed. You must now read the image uri from property `uri` of the returned object. [Here is an example](https://github.com/bamlab/react-native-image-resizer/commit/15ea06d7651faf316b946170427efa90ea48dc4e). Easy, huh?
+`$ react-native link react-native-image-utils`
 
-## Older versions:
+### Manual installation
 
-* 👨 React Native >= 0.29.2 and < 0.40
-```
-npm install --save react-native-image-resizer@0.0.12
-react-native link react-native-image-resizer
-```
 
-* 👴 React Native >= 0.28 and < 0.29.2
-```
-npm install rnpm -g
-rnpm install react-native-image-resizer@0.0.12
-```
+#### iOS
 
-### Android
+1. In XCode, in the project navigator, right click `Libraries` ➜ `Add Files to [your project's name]`
+2. Go to `node_modules` ➜ `react-native-image-utils` ➜ `ios` and add `AEImageUtils.xcodeproj`
+3. In XCode, in the project navigator, select your project. Add `libAEImageUtils.a` to your project's `Build Phases` ➜ `Link Binary With Libraries`
+4. Run your project (`Cmd+R`)<
 
-Note: on latest versions of React Native, you may have an error during the Gradle build on Android (`com.android.dex.DexException: Multiple dex files define Landroid/support/v7/appcompat/R$anim`). Run `cd android && ./gradlew clean` to fix this.
+#### Android
 
-#### Manual linking
-Manual link information for Android: [Link](docs/android_manual_config.md)
+1. Open up `android/app/src/main/java/[...]/MainActivity.java`
+  - Add `import com.appfolio.react.imageutils.AEImageUtilsPackage;` to the imports at the top of the file
+  - Add `new AEImageUtilsPackage()` to the list returned by the `getPackages()` method
+2. Append the following lines to `android/settings.gradle`:
+  	```
+  	include ':react-native-image-utils'
+  	project(':react-native-image-utils').projectDir = new File(rootProject.projectDir, 	'../node_modules/react-native-image-utils/android')
+  	```
+3. Insert the following lines inside the dependencies block in `android/app/build.gradle`:
+  	```
+      compile project(':react-native-image-utils')
+  	```
 
 ## Usage example
 
 ```javascript
-import ImageResizer from 'react-native-image-resizer';
+import ImageUtils from 'react-native-image-utils';
 
-ImageResizer.createResizedImage(imageUri, newWidth, newHeight, compressFormat, quality, rotation, outputPath).then((response) => {
+ImageUtils.createResizedImage(imageUri, newWidth, newHeight, compressFormat, quality, rotation, outputPath).then((response) => {
   // response.uri is the URI of the new image that can now be displayed, uploaded...
   // response.path is the path of the new image
   // response.name is the name of the new image with the extension
@@ -53,7 +53,7 @@ ImageResizer.createResizedImage(imageUri, newWidth, newHeight, compressFormat, q
 
 ### Sample app
 
-A basic, sample app is available in [the `example` folder](https://github.com/bamlab/react-native-image-resizer/tree/master/example). It uses the module to resize a photo from the Camera Roll.
+A basic, sample app is available in [the `example` folder](https://github.com/bamlab/react-native-image-utils/tree/master/example). It uses the module to resize a photo from the Camera Roll.
 
 ## API
 
@@ -70,10 +70,3 @@ compressFormat | Can be either JPEG, PNG or WEBP (android only).
 quality | A number between 0 and 100. Used for the JPEG compression.
 rotation | Rotation to apply to the image, in degrees, for android. On iOS, rotation is limited (and rounded) to multiples of 90 degrees.
 outputPath | The resized image path. If null, resized image will be stored in cache folder. To set outputPath make sure to add option for rotation too (if no rotation is needed, just set it to 0).
-
-## Other open-source modules by the folks at [BAM](http://github.com/bamlab)
-
- * [rn-camera-roll](https://github.com/bamlab/rn-camera-roll)
- * [react-native-numberpicker-dialog](https://github.com/bamlab/react-native-numberpicker-dialog)
- * [react-native-animated-picker](https://github.com/bamlab/react-native-animated-picker)
- * [cordova-plugin-native-routing](https://github.com/bamlab/cordova-plugin-native-routing)
