+++ android/build.gradle
@@ -3,7 +3,6 @@ buildscript {
         jcenter()
     }
 
-
     dependencies {
         classpath 'com.android.tools.build:gradle:2.2.3'
     }
@@ -12,19 +11,17 @@ buildscript {
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.3"
+    compileSdkVersion 23
+    buildToolsVersion "23.0.1"
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 23
+        targetSdkVersion 22
         versionCode 1
         versionName "1.0"
     }
-    buildTypes {
-        release {
-            minifyEnabled false
-        }
+    lintOptions {
+        abortOnError false
     }
 }
