+++ android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="fr.bamlab.rnimageresizer">
+          package="com.appfolio.react.imageutils">
 </manifest>
