+++ android/src/main/java/com/appfolio/react/imageutils/AEImageUtilsPackage.java
@@ -1,4 +1,4 @@
-package fr.bamlab.rnimageresizer;
+package com.appfolio.react.imageutils;
 
 import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.JavaScriptModule;
@@ -6,24 +6,18 @@ import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.uimanager.ViewManager;
 
-import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-/**
- * Created by almouro on 19/11/15.
- */
-public class ImageResizerPackage implements ReactPackage {
+public class AEImageUtilsPackage implements ReactPackage {
     /**
      * @param reactContext react application context that can be used to create modules
      * @return list of native modules to register with the newly created catalyst instance
      */
     @Override
     public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
-        List<NativeModule> modules = new ArrayList<>();
-        modules.add(new ImageResizerModule(reactContext));
-
-        return modules;
+        return Arrays.<NativeModule>asList(new AEImageUtilsModule(reactContext));
     }
 
     /**
