--- docs/android_manual_config.md
@@ -1,20 +1,0 @@
-#### Manual linking
-If your any reason you don want to link this project using 'react-native link', go to settings.gradle and add
-```
-include ':react-native-image-resizer'
-project(':react-native-image-resizer').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-image-resizer/android')
-```
-then go the file that you build the ReactInstance and add the packager to it.
-
-```
-  ReactInstanceManager.Builder builder = ReactInstanceManager.builder()
-                .setApplication(application)
-                .setDefaultHardwareBackBtnHandler(application.getGAMActivity())
-                .setInitialLifecycleState(LifecycleState.RESUMED)
-                .setCurrentActivity((Activity) application.getGAMActivity())
-                .addPackage(new RealmReactPackage())
-                .addPackage(new MainReactPackageWrapper())
-                .addPackage(new StoreReactNativePackage())
-                .addPackage(new ImageResizerPackage()) <------- (Add this package on the Builder list)
-                .addPackage(gamCommunicationReactPackage);
-```
