+++ example/README.md
@@ -1,4 +1,4 @@
-# `react-native-image-resizer` example
+# `react-native-image-utils` example
 
 This is a very simple example app, loading the first photo from the device camera roll
 and allowing to resize it.
@@ -26,11 +26,11 @@ react-native run-android
 
 ## Development
 
-If you want to use this sample app to develop `react-native-image-resizer`, erase the `node_modules/react-native-image-resizer` folder and replace it by the git repository:
+If you want to use this sample app to develop `react-native-image-utils`, erase the `node_modules/react-native-image-utils` folder and replace it by the git repository:
 
 ```bash
-rm -rf node_modules/react-native-image-resizer
-cd node_modules && git clone git@github.com:bamlab/react-native-image-resizer.git
+rm -rf node_modules/react-native-image-utils
+cd node_modules && git clone git@github.com:appfolio/react-native-image-utils.git
 ```
 
 Recompile and relaunch iOS/Android apps.
