+++ example/android/app/build.gradle
@@ -126,7 +126,7 @@ android {
 }
 
 dependencies {
-    compile project(':react-native-image-resizer')
+    compile project(':react-native-image-utils')
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.0.1"
     compile "com.facebook.react:react-native:+"  // From node_modules
