+++ example/android/settings.gradle
@@ -1,5 +1,5 @@
 rootProject.name = 'ResizerExample'
-include ':react-native-image-resizer'
-project(':react-native-image-resizer').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-image-resizer/android')
+include ':react-native-image-utils'
+project(':react-native-image-utils').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-image-utils/android')
 
 include ':app'
