+++ example/package.json
@@ -9,6 +9,6 @@
     "react": "^15.4.0",
     "react-native": "^0.41.0",
     "react-native-gifted-spinner": "^0.1.0",
-    "react-native-image-resizer": "file:../"
+    "react-native-image-utils": "file:../"
   }
 }
