+++ index.android.js
@@ -1,11 +1,11 @@
 import React from 'react-native';
 
-const ImageResizerAndroid = React.NativeModules.ImageResizerAndroid;
+const AEImageUtils = React.NativeModules.AEImageUtils;
 
 export default {
   createResizedImage: (imagePath, newWidth, newHeight, compressFormat, quality, rotation = 0, outputPath) => {
     return new Promise((resolve, reject) => {
-      ImageResizerAndroid.createResizedImage(imagePath, newWidth, newHeight,
+      AEImageUtils.createResizedImage(imagePath, newWidth, newHeight,
         compressFormat, quality, rotation, outputPath, resolve, reject);
     });
   },
