+++ ios/AEImageUtils.h
@@ -1,0 +1,9 @@
+#if __has_include("RCTBridgeModule.h")
+#import "RCTBridgeModule.h"
+#else
+#import <React/RCTBridgeModule.h>
+#endif
+
+@interface AEImageUtils : NSObject <RCTBridgeModule>
+
+@end
