+++ ios/ImageHelpers.h
