+++ ios/ImageHelpers.m
