--- ios/RCTImageResizer/RCTImageResizer.h
@@ -1,5 +1,0 @@
-#import <React/RCTBridgeModule.h>
-
-@interface ImageResizer : NSObject <RCTBridgeModule>
-
-@end
